Addons

+++ build.php
@@ -7,12 +7,23 @@
 declare(strict_types=1);
 set_time_limit(0);
 ini_set("memory_limit", "-1");
-$enable_version_suffix = isset(getopt("vs")["vs"]);
-$secure = getenv("COMPUTERNAME") !== "JANPC";
-$buildOnLocalServer = false;
+
+///////////////////////////////////////////////////////////////////////////////////////////////////
+///////////////////////////////////////////////////////////////////////////////////////////////////
+///////////////////////////////////////////////////////////////////////////////////////////////////
+$localServerPath = "C:/Users/" . getenv("USERNAME") . "/Desktop/pmmp4"; // string|null
+$NAMESPACE = "HQGames/Core";
 $packages = [
-	//EXAMPLE: "xxarox/web-server": ["paths" => ["src/","README.md"], "encode" => true]
+	"xxarox/xxtools" => ["paths" => ["src/xxAROX/xxTOOLS" => "src/xxAROX/xxTOOLS"], "encode" => false]
 ];
+$encode = getenv("COMPUTERNAME") !== "JANPC";
+$enable_version_suffix = false;
+///////////////////////////////////////////////////////////////////////////////////////////////////
+///////////////////////////////////////////////////////////////////////////////////////////////////
+///////////////////////////////////////////////////////////////////////////////////////////////////
+echo "[INFO]: Updating autoloader" . PHP_EOL;
+passthru("composer  --no-dev --no-interaction dump-autoload -o", $result_code);
+if ($result_code != 0) throw new ErrorException("Error while updated autoloader.");
 $loader = include_once __DIR__ . "/vendor/autoload.php";
 $startTime = microtime(true);
 $from = getcwd() . DIRECTORY_SEPARATOR;
@@ -22,73 +33,48 @@ $outputPath = $from . "out" . DIRECTORY_SEPARATOR . $description["name"] . ($ena
 echo "[INFO]: Starting.." . PHP_EOL;
 @mkdir($to, 0777, true);
 cleanDirectory($to);
+
+// include all important files
+if (is_dir($from . "src")) copyDirectory($from . "src", $to . "src/${NAMESPACE}");
+if (is_file($from . "LICENSE")) file_put_contents($to . "LICENSE", file_get_contents($from . "LICENSE"));
+if (is_file($from . "README.md")) file_put_contents($to . "README.md", file_get_contents($from . "README.md"));
+if (is_dir($from . "resources")) copyDirectory($from . "resources", $to . "resources");
+yaml_emit_file($to . "plugin.yml", $description);
+
+// include all packages
 foreach ($packages as $vendor => $obj) {
-	if (str_ends_with($vendor, "/")) {
-		$vendor = substr($vendor, 0, -1);
-	}
-	foreach ($obj["paths"] as $path) {
+	if (str_ends_with($vendor, "/")) $vendor = substr($vendor, 0, -1);
+	foreach ($obj["paths"] as $path => $better_path) {
 		if (is_dir($from . "vendor/$vendor")) {
-			$package = json_decode(file_get_contents(__DIR__ . "vendor/{$vendor}/composer.json"), true);
-			foreach ($package["autoload"] ?? [] as $type => $obj) {
-				if (isset($package["autoload"][$type]) && is_string($package["autoload"][$type])) {
-					$package["autoload"][$type] = [$package["autoload"][$type]];
-					foreach ($package["autoload"][$type] as $k => $file) {
-						copyDirectory($from . "vendor/$vendor/$path/$file", "$vendor/$path/$file");
-						$package["autoload"][$type][$k] = "$vendor/$path/$file";
-					}
-				}
-			}
-			file_put_contents($to . "vendor/{$vendor}/package.json", json_encode($package, JSON_PRETTY_PRINT));
+			copyDirectory($from . "vendor/$vendor/$path", $to . $better_path);
 		} else {
 			throw new RuntimeException("Package '$vendor' is not installed.");
 		}
 	}
 }
-echo "[INFO]: Loaded " . count($packages) . " packages" . PHP_EOL;
-if (is_dir($from . "src")) {
-	copyDirectory($from . "src", $to . "src/HQGames/Core");
-}
-if (is_file($from . "LICENSE")) {
-	file_put_contents($to . "LICENSE", file_get_contents($from . "LICENSE"));
-}
-if (is_file($from . "README.md")) {
-	file_put_contents($to . "README.md", file_get_contents($from . "README.md"));
-}
-if (is_dir($from . "resources")) {
-	copyDirectory($from . "resources", $to . "resources");
-}
-yaml_emit_file($to . "plugin.yml", $description);
-echo "[INFO]: Plugin files files" . PHP_EOL;
-passthru("composer  --no-dev --no-interaction dump-autoload -o", $result_code);
-if ($result_code != 0) {
-	throw new ErrorException("Error while updated autoloader.");
-}
+echo "[INFO]: Included " . count($packages) . " package" . (count($packages) == 1 ? "" : "s") . PHP_EOL;
+
+// plugin encoder
 $excluded = [];
 foreach ($packages as $vendor => $obj) {
-	if ($obj["encode"] ?? false) {
-		$excluded[] = $vendor . "/";
-	}
+	if (!($obj["encode"] ?? false)) $excluded[] = $vendor;
 }
-if ($secure) {
+echo "[INFO]: Encoding plugin.." . PHP_EOL;
+if ($encode) {
 	echo "[INFO]: Encoding plugin.." . PHP_EOL;
 	require_once "vendor/xxarox/plugin-security/src/Encoder.php";
 	(new \xxAROX\PluginSecurity\Encoder($to, $excluded))->encode();
 	echo "[INFO]: Encoding done!" . PHP_EOL;
+	$to = $to . "/output";
 }
-if (is_dir($to . "output/")) {
-	$to = $to . "output/";
-}
+echo "[INFO]: Encoding done!" . PHP_EOL;
 generatePhar($outputPath, $to);
-if ($buildOnLocalServer && is_dir("C:/Users/" . getenv("USERNAME") . "/Desktop/pmmp" . (is_array($description["api"])
-			? explode(".", $description["api"][0])[0]
-			: (is_string($description["api"]) ? explode(".", $description["api"])[0] : "???")) . "/plugins")) {
-	echo "[INFO]: Building on " . PHP_EOL;
-	$outputPath = "C:/Users/" . getenv("USERNAME") . "/Desktop/pmmp" . (is_array($description["api"])
-			? explode(".", $description["api"][0])[0]
-			: (is_string($description["api"]) ? explode(".", $description["api"])[0]
-				: "???")) . "/plugins" . DIRECTORY_SEPARATOR . $description["name"] . ($enable_version_suffix ? "_v" . $description["version"] : "");
-	generatePhar($outputPath, $to);
+
+if (!empty($localServerPath) && is_dir($localServerPath . "/plugins")) {
+	echo "[INFO]: Compiling.." . PHP_EOL;
+	generatePhar($localServerPath . "/plugins/" . $description["name"] . ($enable_version_suffix ? "_v" . $description["version"] : ""), $to);
 }
+
 /**
  * Function copyDirectory
  * @param string $from
